backend/requirements.txt
# Core
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.2
python-multipart>=0.0.6

# LLM & AI
langchain>=0.1.0
langchain-google-genai>=0.0.10
google-generativeai>=0.3.0

# File Processing
pypdf2>=3.0.0
pdfplumber>=0.10.3
python-docx>=1.0.0
Pillow>=10.0.0
pdf2image>=1.16.3
pytesseract>=0.3.10

# Utils
python-magic>=0.4.27
python-magic-bin>=0.4.14; sys_platform == 'win32'

# Development
pytest>=7.4.0
black>=23.7.0
isort>=5.12.0
mypy>=1.5.1
